fix of accessing PROGMEM strings above 64k flash.
The avr-gcc 4.7.2 complains about the wrong pointer size of PGM_P! So I
switched to uint_farptr_t for the Mega-128. Since pointers are always 16bit,
it isn't possible to get the correct address by simply casting the pointer.
I've found the macro FAR(), which restaurates a 17bit address. The drawback of
the macro is, that is only works with variables. It is not possible to pass a
PSTR() constant to it. As a simple solution, I've put all messages into
variables.